+++ src/programs/AmplitudeAnalysis/project_moments_polarized/project_moments_polarized.cpp
@@ -105,8 +105,8 @@ positive.push_back(wave("G1+", 4, 1));
   Biggest_lm(ws, &LMAX);
 
 
-  double step = ( highMass - lowMass ) / kNumBins;
-    double stept = ( hight - lowt ) / kNumBinst;
+    double step = ( highMass - lowMass ) / static_cast<double>(kNumBins);
+    double stept = ( hight - lowt ) / static_cast<double>(kNumBinst);
     
     ofstream outfile;
     outfile.open( outfileName.c_str() );
